move QueryGenerator to Query sub-namepace
